refactor(frontend)[OK-325] Move errors and loading components to appropriate routes
