+++ client/src/app/user-boards/(board-management)/[boardId]/_components/board-details-sidebar/BoardDetailsSidebarContent.tsx
