+++ client/src/app/user-boards/(board-management)/[boardId]/_loading/LoadingSidebar.tsx
