+++ client/src/app/user-boards/(board-management)/[boardId]/permissions/_components/BoardPermissions.tsx
@@ -3,14 +3,14 @@ import { BoardPermissionsUser } from "@/interfaces/board-permissions-user";
 import React from "react";
 import { BoardPermission } from "@/enums/BoardPermission";
 import { SectionHeader } from "@/components/common/headers/section-header/SectionHeader";
-import BoardPermissionsInfoDialog from "@/components/common/dialogs/BoardPermissionsInfoDialog";
+import BoardPermissionsInfoDialog from "@/app/user-boards/(board-management)/[boardId]/permissions/_components/BoardPermissionsInfoDialog";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
 import { Share2 } from "lucide-react";
 import { Table, TableBody, TableCell, TableRow } from "@/components/ui/table";
 import BoardPermissionsSelect from "@/components/common/board-permissions-select/BoardPermissionsSelect";
 import { BoardTableLeadRow } from "@/app/user-boards/(board-management)/_components/board-table/BoardTableLeadRow";
-import DeleteCollaboratorDialog from "@/components/common/dialogs/DeleteCollaboratorDialog";
+import DeleteCollaboratorDialog from "@/app/user-boards/(board-management)/[boardId]/permissions/_components/DeleteCollaboratorDialog";
 import { BoardDetailsResponse } from "@/interfaces/responses/board-details-response";
 import ShareBoardDialog from "@/components/common/dialogs/ShareBoardDialog";
 import { boardPermissionToNum, capabilityFunctions } from "@/lib/permissionUtils";
