+++ client/src/app/user-boards/(board-management)/[boardId]/permissions/_components/DeleteCollaboratorDialog.tsx
