+++ client/src/app/user-boards/(board-management)/_components/board-table/BoardDetailsButton.tsx
