+++ client/src/app/user-boards/(board-management)/_components/board-table/BoardTable.tsx
@@ -1,7 +1,7 @@
 import { FilterIcon, Share2, ViewIcon } from "lucide-react";
 import React, { useCallback } from "react";
 import { Table, TableBody, TableCell, TableRow } from "@/components/ui/table";
-import { UserBoardsPagination } from "../../../../../components/common/paginations/UserBoardsPagination";
+import { UserBoardsPagination } from "./UserBoardsPagination";
 import { getColumnsForTab, getDescriptionForTab, getTitleForTab } from "@/lib/userBoardsUtils";
 import { Button } from "@/components/ui/button";
 import { CellContent } from "@/app/user-boards/(board-management)/_components/board-table/CellContent";
@@ -12,8 +12,8 @@ import { BoardsActiveTab, BoardsActiveTabMap } from "@/enums/BoardsActiveTab";
 import { SectionHeader } from "@/components/common/headers/section-header/SectionHeader";
 import { BoardTableLeadRow } from "@/app/user-boards/(board-management)/_components/board-table/BoardTableLeadRow";
 import { useRouter } from "next/navigation";
-import { DeleteBoardDialog } from "@/components/common/dialogs/DeleteBoardDialog";
-import { BoardDetailsButton } from "@/components/common/buttons/BoardDetailsButton";
+import { DeleteBoardDialog } from "@/app/user-boards/(board-management)/_components/board-table/DeleteBoardDialog";
+import { BoardDetailsButton } from "@/app/user-boards/(board-management)/_components/board-table/BoardDetailsButton";
 import { deleteBoard } from "@/app/actions/boardActions";
 import ShareBoardDialog from "@/components/common/dialogs/ShareBoardDialog";
 import { capabilityFunctions } from "@/lib/permissionUtils";
