+++ client/src/app/user-boards/(realtime-canvas)/[boardId]/slides/[slideIndex]/_components/board-sidebar/BoardSidebarItem.tsx
