+++ client/src/app/user-boards/(realtime-canvas)/[boardId]/slides/[slideIndex]/_components/slide-controls/DeleteSlideDialog.tsx
