+++ client/src/app/user-boards/(realtime-canvas)/[boardId]/slides/[slideIndex]/_error/BoardError.tsx
