+++ client/src/interfaces/error-component/error-component.ts
@@ -1,0 +1,4 @@
+export interface ErrorComponentProps {
+  error: Error & { digest?: string };
+  reset: () => void;
+}
